+++ contracts/capability-contract/src/index.ts
@@ -16,3 +16,5 @@ export * from "./fake-host-adapter.js";
 
 export * from "./compatibility-types.js";
 export * from "./compatibility.js";
+export * from "./challenge-types.js";
+export * from "./challenge.js";